Update properties for local development
This was using properties for an environment in RedCentric which no longer exists.

Update the properties to use OPA locally. This assumes you're running OPA locally in the same
way that you would for PUI local development (in a Docker container). There is more details on
how to set this up in PUI README here https://github.com/ministryofjustice/laa-ccms-pui/blob/master/documentation/gradle-docker-build.md

Also add the property for the billing address.