Remove unused maven profiles and properties files
Now that we're using Docker compose we can pass in the spring profile and other properties as an environment variable
at run time, which means we don't need to use Maven profiles or have a separate properties file to set them.

We've left Stage as it is for now in case we need to re-build this adapter to use it in Staging for testing before
we migrate to AWS. To do that you'd also need to uncomment `<!--<packaging>war</packaging>-->` because we need to build
a WAR instead of a fat JAR to deploy to Weblogic.